requirements.txt
openai
langchain
streamlit
python-dotenv
langchain_community
pyPDF2
faiss-cpu
tiktoken